better cache config part 1
use a list of caches which is “above” the metrics writers in hierarchy as for multi writers we need a good way to specifiy the cache to use (as we can just use one if the data sets are comensurate).

If using multi writers, one series and one not-series (i.e. cassandra + whisper) one will need multiple caches as the two have very different writer patterns.

This also les us specifiy in the api config which cache to use for the readers w/o having weird options not related to readers