WIP: a new series writer type in the works .. Log
This type uses another table to drop a “blob” all the metrics in a given 10s window into a log (current just another cassandra table).  Rather then maintain a single list based on overflow (where once a metric hits a byte size, it will get written) this one is Time chunk based.  Where N chunks are kept for smaller time slices, once a chunk is written the SequenceID in the Log tables are “removed”.  So upon a crash, the system just grabs and re-makes the chunks from what’s currently in that log table.  So hopefully at most only the Log Time chunk is potentially lost alot of metrics that were in ram.